+++ CHANGELOG.md
@@ -2,6 +2,7 @@
   - Auto-generates "All Tracks Playlist"
   - Hotspot password/passcode merged
   - Calculate Track time remaining
+  - Streaming coordinates
   - Fix Ansible reconnect
 # 2020-03-11 1.10.67-70 504acb2722c107f9898ad105bd6dd671b5e8958f
   - Clear passcode on GPIO hold