+++ CHANGELOG.md
@@ -1,4 +1,7 @@
-# 2020-05-22 1.10.71-72
+# 2020-08-05 1.10.73
+  - System stats endpoint
+  - State added to LEDs
+# 2020-05-22 1.10.71-72 1f9746343c83bb5c388fc85eeac234ba5763cd1b
   - Auto-generates "All Tracks Playlist"
   - Hotspot password/passcode merged
   - Calculate Track time remaining